+++ plugin-dev/Source/Sentry/Private/GenericPlatform/GenericPlatformSentrySpan.h
@@ -25,7 +25,7 @@ public:
 	virtual bool IsFinished() const override;
 	virtual void SetTag(const FString& key, const FString& value) override;
 	virtual void RemoveTag(const FString& key) override;
-	virtual void SetData(const FString& key, const TMap<FString, FString>& values) override;
+	virtual void SetData(const FString& key, const TMap<FString, FSentryVariant>& values) override;
 	virtual void RemoveData(const FString& key) override;
 	virtual void GetTrace(FString& name, FString& value) override;