+++ plugin-dev/Source/Sentry/Private/SentryOutputDevice.cpp
@@ -41,7 +41,7 @@ void FSentryOutputDevice::Serialize(const TCHAR* V, ELogVerbosity::Type Verbosit
 		return;
 	}
 
-	SentrySubsystem->AddBreadcrumbWithParams(Message, Category.ToString(), FString(), TMap<FString, FString>(), BreadcrumbLevel);
+	SentrySubsystem->AddBreadcrumbWithParams(Message, Category.ToString(), FString(), TMap<FString, FSentryVariant>(), BreadcrumbLevel);
 }
 
 bool FSentryOutputDevice::CanBeUsedOnAnyThread() const