+++ plugin-dev/Source/Sentry/Private/Tests/SentryBreadcrumb.spec.cpp
@@ -28,9 +28,11 @@ void SentryBreadcrumbSpec::Define()
 			const FString TestType = FString(TEXT("Test type"));
 			const FString TestCategory = FString(TEXT("Test category"));
 
-			TMap<FString, FString> TestData;
+			TMap<FString, FSentryVariant> TestData;
 			TestData.Add(TEXT("Key1"), TEXT("Val1"));
-			TestData.Add(TEXT("Key2"), TEXT("Val2"));
+			TestData.Add(TEXT("Key2"), 123);
+			TestData.Add(TEXT("Key3"), 4.5f);
+			TestData.Add(TEXT("Key4"), true);
 
 			SentryBreadcrumb->SetLevel(ESentryLevel::Fatal);
 			SentryBreadcrumb->SetMessage(TestMessage);
@@ -43,9 +45,11 @@ void SentryBreadcrumbSpec::Define()
 			TestEqual("Breadcrumb type", SentryBreadcrumb->GetType(), TestType);
 			TestEqual("Breadcrumb category", SentryBreadcrumb->GetCategory(), TestCategory);
 
-			TMap<FString, FString> ReceivedData = SentryBreadcrumb->GetData();
+			TMap<FString, FSentryVariant> ReceivedData = SentryBreadcrumb->GetData();
 			TestEqual("Data 1", ReceivedData[TEXT("Key1")], TestData[TEXT("Key1")]);
 			TestEqual("Data 2", ReceivedData[TEXT("Key2")], TestData[TEXT("Key2")]);
+			TestEqual("Data 3", ReceivedData[TEXT("Key3")], TestData[TEXT("Key3")]);
+			TestEqual("Data 4", ReceivedData[TEXT("Key4")], TestData[TEXT("Key4")]);
 		});
 	});
 }