+++ plugin-dev/Source/Sentry/Public/SentryTransaction.h
@@ -6,6 +6,7 @@
 #include "UObject/Object.h"
 
 #include "SentryImplWrapper.h"
+#include "SentryVariant.h"
 
 #include "SentryTransaction.generated.h"
 
@@ -58,7 +59,7 @@ public:
 
 	/** Sets data associated with the transaction. */
 	UFUNCTION(BlueprintCallable, Category = "Sentry")
-	void SetData(const FString& key, const TMap<FString, FString>& values);
+	void SetData(const FString& key, const TMap<FString, FSentryVariant>& values);
 
 	/** Removes data associated with the transaction. */
 	UFUNCTION(BlueprintCallable, Category = "Sentry")